civil-interview-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
pandas
openpyxl
pydantic
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy
pymysql
cryptography
python-dotenv
